+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/SceneManager.kt
@@ -1,3 +1,5 @@
+@file:Suppress("unused")
+
 package ru.cherryngine.engine.scenes
 
 import io.micronaut.context.ApplicationContext
@@ -8,10 +10,12 @@ import org.jgrapht.Graph
 import org.jgrapht.graph.DefaultEdge
 import org.jgrapht.graph.DirectedAcyclicGraph
 import org.jgrapht.traverse.TopologicalOrderIterator
+import ru.cherryngine.engine.scenes.api.Module
+import ru.cherryngine.engine.scenes.api.ModulePrototype
 import ru.cherryngine.engine.scenes.event.Event
 import java.util.*
 import kotlin.reflect.KClass
-import kotlin.reflect.full.*
+import kotlin.reflect.full.findAnnotation
 
 @Singleton
 class SceneManager(
