+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/modules/Synchronizer.kt
@@ -1,5 +1,5 @@
 package ru.cherryngine.engine.scenes.modules
 
-import ru.cherryngine.engine.scenes.Module
+import ru.cherryngine.engine.scenes.api.Module
 
 interface Synchronizer : Module
