+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/modules/TransformModule.kt
@@ -2,8 +2,8 @@ package ru.cherryngine.engine.scenes.modules
 
 import io.micronaut.context.annotation.Parameter
 import ru.cherryngine.engine.scenes.GameObject
-import ru.cherryngine.engine.scenes.Module
-import ru.cherryngine.engine.scenes.ModulePrototype
+import ru.cherryngine.engine.scenes.api.Module
+import ru.cherryngine.engine.scenes.api.ModulePrototype
 import ru.cherryngine.lib.math.Transform
 import ru.cherryngine.lib.math.Vec3D
 import ru.cherryngine.lib.math.rotation.QRot
