--- engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/modules/client/Camera.kt
@@ -1,5 +1,0 @@
-package ru.cherryngine.engine.scenes.modules.client
-
-import ru.cherryngine.engine.scenes.Module
-
-interface Camera : Module
