--- engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/modules/client/Controller.kt
@@ -1,7 +1,0 @@
-package ru.cherryngine.engine.scenes.modules.client
-
-import ru.cherryngine.engine.scenes.Module
-import ru.cherryngine.lib.math.Vec3D
-import ru.cherryngine.lib.math.rotation.QRot
-
-interface Controller : Module
