+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/modules/client/api/Camera.kt
@@ -1,0 +1,5 @@
+package ru.cherryngine.engine.scenes.modules.client.api
+
+import ru.cherryngine.engine.scenes.api.Module
+
+interface Camera : Module
