+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/view/Viewable.kt
@@ -1,6 +1,6 @@
 package ru.cherryngine.engine.scenes.view
 
-import ru.cherryngine.engine.scenes.Module
+import ru.cherryngine.engine.scenes.api.Module
 
 interface Viewable : Module {
     fun showFor(viewer: Viewer): Boolean
