+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/view/Viewer.kt
@@ -1,8 +1,8 @@
 package ru.cherryngine.engine.scenes.view
 
-import ru.cherryngine.engine.scenes.Module
+import ru.cherryngine.engine.scenes.api.Module
 
 interface Viewer : Module {
-    fun show(viewable: Viewable): Boolean
-    fun hide(viewable: Viewable): Boolean
+    fun show(viewable: Viewable): Boolean { return true }
+    fun hide(viewable: Viewable): Boolean { return true }
 }
