+++ impl-demo/src/main/kotlin/ru/cherryngine/impl/demo/Projectile.kt
@@ -8,12 +8,11 @@ import net.minestom.server.item.Material
 import ru.cherryngine.engine.core.minestomPos
 import ru.cherryngine.engine.core.world.entity.EngineEntity
 import ru.cherryngine.engine.scenes.GameObject
-import ru.cherryngine.engine.scenes.Module
-import ru.cherryngine.engine.scenes.ModulePrototype
 import ru.cherryngine.engine.scenes.Scene
+import ru.cherryngine.engine.scenes.api.ModulePrototype
 import ru.cherryngine.engine.scenes.event.Event
 import ru.cherryngine.engine.scenes.modules.client.ClientModule
-import ru.cherryngine.engine.scenes.modules.physics.collider.Collider
+import ru.cherryngine.engine.scenes.modules.physics.api.Collider
 import ru.cherryngine.engine.scenes.view.Viewable
 import ru.cherryngine.engine.scenes.view.Viewer
